requirements.txt
jsonpath==0.82
requests==2.24.0
python-dotenv~=1.0.1
pytest==7.4.0
pytest-html==4.1.1
py==1.8.2
pytest_steps==1.8.0
pytest-html-reporter==0.2.9